Split out service
The only user of the service is `Room`, so we've moved the service into its own file, and only refer to it from the `Room` class.
